+++ app/src/main/java/de/baumann/browser/browser/NinjaDownloadListener.java
@@ -8,11 +8,15 @@ import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Environment;
+import android.os.Handler;
+import android.os.HandlerThread;
+import android.os.Message;
 import android.view.Gravity;
 import android.view.View;
 import android.webkit.CookieManager;
 import android.webkit.DownloadListener;
 import android.webkit.URLUtil;
+import android.webkit.WebView;
 import android.widget.GridView;
 import android.widget.Toast;
 
@@ -35,22 +39,34 @@ import de.baumann.browser.view.GridItem;
 
 public class NinjaDownloadListener implements DownloadListener {
     private final Context context;
+    private final WebView webView;
 
-    public NinjaDownloadListener(Context context) {
+    public NinjaDownloadListener(Context context, WebView webView) {
         super();
         this.context = context;
+        this.webView = webView;
+
     }
 
     @Override
     public void onDownloadStart(final String url, String userAgent, final String contentDisposition, final String mimeType, long contentLength) {
 
-        String filename = URLUtil.guessFileName(url, contentDisposition, mimeType);
+        // Create a background thread that has a Looper
+        HandlerThread handlerThread = new HandlerThread("HandlerThread");
+        handlerThread.start();
+        // Create a handler to execute tasks in the background thread.
+        Handler backgroundHandler = new Handler(handlerThread.getLooper());
+        Message msg = backgroundHandler.obtainMessage();
+        webView.requestFocusNodeHref(msg);
+        final String[] msgString = {(String) msg.getData().get("url")};
+
+        String filename = URLUtil.guessFileName(msgString[0], null, null);
         String text = context.getString(R.string.dialog_title_download) + " - " + filename;
 
-        GridItem item_01 = new GridItem(context.getString(R.string.app_ok), 0);
-        GridItem item_02 = new GridItem( context.getString(R.string.menu_share_link), 0);
-        GridItem item_03 = new GridItem( context.getString(R.string.menu_save_as), 0);
-        GridItem item_04 = new GridItem( context.getString(R.string.app_cancel), 0);
+        GridItem item_01 = new GridItem(context.getString(R.string.app_ok), R.drawable.icon_check);
+        GridItem item_02 = new GridItem( context.getString(R.string.menu_share_link), R.drawable.icon_link);
+        GridItem item_03 = new GridItem( context.getString(R.string.menu_save_as), R.drawable.icon_save_as);
+        GridItem item_04 = new GridItem( context.getString(R.string.app_cancel), R.drawable.icon_close);
 
         MaterialAlertDialogBuilder builder = new MaterialAlertDialogBuilder(context);
         builder.setTitle(R.string.app_warning);
@@ -81,18 +97,22 @@ public class NinjaDownloadListener implements DownloadListener {
                 case 0:
                     try {
                         Activity activity = (Activity) context;
-                        if (url.startsWith("data:")) {
-                            DataURIParser dataURIParser = new DataURIParser(url);
+                        if (msgString[0] == null) {
+                            msgString[0] = url;
+                        }
+                        assert msgString[0] != null;
+                        if (msgString[0].startsWith("data:")) {
+                            DataURIParser dataURIParser = new DataURIParser(msgString[0]);
                             if (BackupUnit.checkPermissionStorage(context)) {
                                 File file = new File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS), filename);
                                 FileOutputStream fos = new FileOutputStream(file);
                                 fos.write(dataURIParser.getImagedata()); }
                             else BackupUnit.requestPermission(activity); }
                         else {
-                            DownloadManager.Request request = new DownloadManager.Request(Uri.parse(url));
+                            DownloadManager.Request request = new DownloadManager.Request(Uri.parse(msgString[0]));
                             request.setMimeType(mimeType);
                             //------------------------COOKIE!!------------------------
-                            String cookies = CookieManager.getInstance().getCookie(url);
+                            String cookies = CookieManager.getInstance().getCookie(msgString[0]);
                             request.addRequestHeader("cookie", cookies);
                             //------------------------COOKIE!!------------------------
                             request.setDescription(context.getString(R.string.dialog_title_download));
@@ -113,7 +133,7 @@ public class NinjaDownloadListener implements DownloadListener {
                     try {
                         Intent sharingIntent = new Intent(Intent.ACTION_SEND);
                         sharingIntent.setType("text/plain");
-                        sharingIntent.putExtra(Intent.EXTRA_TEXT, url);
+                        sharingIntent.putExtra(Intent.EXTRA_TEXT, msgString[0]);
                         context.startActivity(Intent.createChooser(sharingIntent, (context.getString(R.string.menu_share_link)))); }
                     catch (Exception e) {
                         System.out.println("Error Downloading File: " + e);
@@ -122,7 +142,7 @@ public class NinjaDownloadListener implements DownloadListener {
                     break;
                 case 2:
                     Activity activity = (Activity) context;
-                    HelperUnit.saveAs(activity, url, filename, dialog);
+                    HelperUnit.saveAs(activity, msgString[0], filename, dialog);
                     break;
                 case 3:
                     break;